+++ src/memoConfigure.ts
@@ -55,6 +55,7 @@ export class memoConfigure {
     public memoAdminOpenMode: string;
     public memoAdminOpenOnStartup: boolean;
     public memoAdminRecentTitleMode: string;
+    public memoPinnedFiles: string[];
     public openMarkdownPreviewUseMPE: boolean;
     public memoOpenChromeCustomizeURL: string;
     public memoTyporaExecPath: string;
@@ -178,6 +179,7 @@ export class memoConfigure {
         this.memoAdminOpenMode = vscode.workspace.getConfiguration('memo-life-for-you').get<string>('memoAdminOpenMode');
         this.memoAdminOpenOnStartup = vscode.workspace.getConfiguration('memo-life-for-you').get<boolean>('memoAdminOpenOnStartup');
         this.memoAdminRecentTitleMode = vscode.workspace.getConfiguration('memo-life-for-you').get<string>('memoAdminRecentTitleMode');
+        this.memoPinnedFiles = vscode.workspace.getConfiguration('memo-life-for-you').get<string[]>('memoPinnedFiles');
         this.openMarkdownPreviewUseMPE = vscode.workspace.getConfiguration('memo-life-for-you').get<boolean>('openMarkdownPreviewUseMPE');
         this.memoOpenChromeCustomizeURL = vscode.workspace.getConfiguration('memo-life-for-you').get<string>('openChromeCustomizeURL');
         this.memoTyporaExecPath = vscode.workspace.getConfiguration('memo-life-for-you').get<string>('TyporaExecPath');